backend/requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.2
Django==5.1.6
django-cors-headers==4.7.0
django-email==0.1.10
django-rest-knox==4.2.0
djangorestframework==3.15.1
pycparser==2.22
sqlparse==0.5.3
tzdata==2025.1